Extract token from lexer
